wgengine/magicsock: clear out endpoint statistics when it becomes bad
There are cases where we do not detect the non-viability of a route, but
we will instead observe a failure to send. In a Disco path this would
normally be handled as a side effect of Disco, which is not available to
non-Disco WireGuard nodes. In both cases, recognizing the failure as
such will result in faster convergence.

Updates #8999

Cherry picked from 9f515676ce90c6bb438afbf7ca2b4be9f2a98c24

Signed-off-by: James Tucker <james@tailscale.com>