cmd/tailscale,ipn,types/preftype,wgengine/router: introduce netfilter modes representing new capabilities
These are not yet plumbed to work, and are not yet documented in the CLI
help output. If accepted they would take the place of the debug flags,
and provide a persistent way for users to explicitly switch firewall
modes between the integrations that we support (currently iptables and
nftables).

Signed-off-by: James Tucker <james@tailscale.com>