.github/workflows: disable generate check for util/winutil
util/winutil is performing a two-stage generate first generating syscall
stubs, one with an external package reference, and then using goimports
that appears to have previously imported that package. As of current
main, this is not working, and so flows that trigger the go_generate
check will fail on non-matching output.

Updates #9579
Signed-off-by: James Tucker <james@tailscale.com>