net/dns/resolver, util/race: factor out race code into package
This adds a new package, util/race, to contain the logic required for
racing two different functions and returning the first non-error answer.

The DNS forwarder code is responsible for signaling a retry by
converting a successful but truncated response into an error in the UDP
query case, which triggers the TCP fallback immediately.

Signed-off-by: Andrew Dunham <andrew@du.nham.ca>
Change-Id: Ide33e4bc225158c598444da524952880e34ff0dd