cmd/containerboot: revert to using tailscale up
This partially reverts commits a61a9ab087e16270bc039252e7620aae4de3d56e
and 7538f386710b80c6b4c1997797be28a661210d4a and fully reverts
4823a7e591ef859250114ad20b337d4358af9ead.

The goal of that commit was to reapply known config whenever the
container restarts. However, that already happens when TS_AUTH_ONCE was
false (the default back then). So we only had to selectively reapply the
config if TS_AUTH_ONCE is true, this does exactly that.

This is a little sad that we have to revert to `tailscale up`, but it
fixes the backwards incompatibility problem.

Updates tailscale/tailscale#9539

Signed-off-by: Maisem Ali <maisem@tailscale.com>
Signed-off-by: Alex Paguis <alex@windscribe.com>